Fix: added FCLOSE_ON_EXIT and streamlined _ON_EXIT family
