+++ mdstcpip/ProcessMessage.c
@@ -22,7 +22,6 @@ CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
 OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
-#define DEF_FREEXD
 #if defined(linux) && !defined(_LARGEFILE_SOURCE)
 # define _LARGEFILE_SOURCE
 # define _FILE_OFFSET_BITS 64
