+++ tdishr/TdiDoTask.c
@@ -36,8 +36,6 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
         Ken Klare, LANL P-4     (c)1992
 */
-#define DEF_FREED
-#define DEF_FREEXD
 #include <STATICdef.h>
 #include <stdlib.h>
 #include <mdsdescrip.h>
